Fixes logic around multi blocks of same type and adds tests
Prevously we had 1 function attempting to test for multiple block selection and checking that the selection was all of the same block type.

This caused bugs within `switchToBlockType` because the logic was confusing. For example, if a selection isn’t a multi block then we don’t need to test that all the blocks are the same.

Separated the two functions and updated conditions in switchToBlockType to reflect this.

Added unit tests to cover two new functions.